+++ src/app/index.tsx
@@ -80,7 +80,7 @@ export default function SplashScreen() {
         }}
       >
         <Image
-          source={require('../../assets/images/logo.jpg')}
+          source={require('../../assets/images/logo.png')}
           style={styles.logo}
           resizeMode="contain"
           accessible