+++ src/app/modos.tsx
@@ -23,7 +23,7 @@ export default function ModosScreen() {
       </Pressable>
 
       <Image
-        source={require('../../assets/images/logo.jpg')}
+        source={require('../../assets/images/logo.png')}
         style={styles.logo}
         resizeMode="contain"
       />